Update test config to run each test with a fresh db
Rather than manually drop and create databases, we've updated the
test framework to do it for us.  Since the DB URI is loaded from
an environment variable, keeping test and prod DBs separate should
be trival.

If the tester is using a SQLite db, we can simply remove the test
db from the filesystem to clean up after ourselves.  This could
likely be improved to add handing for other DBMS options.